+++ Game/Objs/Obj_Effect_Biomass.cs
@@ -32,7 +32,7 @@ namespace Somnium.Game {
 		// Function from file: biomass.dm
 		public override bool ex_act( double? severity = null, dynamic child = null ) {
 			
-			switch ((double?)( severity )) {
+			switch ((int?)( severity )) {
 				case 1:
 					GlobalFuncs.qdel( this );
 					break;
