+++ Game/Objs/Obj_Item_Device_Aicard.cs
@@ -25,8 +25,8 @@ namespace Somnium.Game {
 
 		// Function from file: aicard.dm
 		public override bool ex_act( double? severity = null, dynamic child = null ) {
-			
-			switch ((double?)( severity )) {
+
+			switch ((int?)(severity)) {
 				case 1:
 					GlobalFuncs.qdel( this );
 					break;
