+++ Game/Objs/Obj_Item_Pipe.cs
@@ -453,8 +453,8 @@ namespace Somnium.Game {
 
 		// Function from file: construction.dm
 		public override bool ex_act( double? severity = null, dynamic child = null ) {
-			
-			switch ((double?)( severity )) {
+
+			switch ((int?)(severity)) {
 				case 1:
 					GlobalFuncs.returnToPool( this );
 					break;
