+++ Game/Objs/Obj_Structure_Plasticflaps.cs
@@ -31,8 +31,8 @@ namespace Somnium.Game {
 
 		// Function from file: supplyshuttle.dm
 		public override bool ex_act( double? severity = null, dynamic child = null ) {
-			
-			switch ((double?)( severity )) {
+
+			switch ((int?)(severity)) {
 				case 1:
 					GlobalFuncs.qdel( this );
 					break;
