+++ Game/Tiles/Tile_Unsimulated_Floor.cs
@@ -36,8 +36,8 @@ namespace Somnium.Game {
 
 		// Function from file: floor.dm
 		public override bool ex_act( double? severity = null, dynamic child = null ) {
-			
-			switch ((double?)( severity )) {
+
+			switch ((int?)(severity)) {
 				case 1:
 					new Obj_Effect_Decal_Cleanable_Soot( this );
 					break;
